Platform/ARM/Sgi: allow MM_STANDALONE modules to use NorFlashPlatformLib
“NorFlashPlatformLib” library can be used by MM_STANDALONE drivers as
well. So add MM_STANDALONE as the module type this library supports.

When used in MM mode, the third instance of the NOR flash is used as
the non-volatile storage. This NOR flash instance is partitioned into
two regions - first 4MB space is used for secure boot and next 3MB for
secure variable storage

Change-Id: I2bbdb735cfdfba5cb64fe0077b2f5d287370a65f
Signed-off-by: Jagadeesh Ujja <jagadeesh.ujja@arm.com>
Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>